+++ app/src/main/java/it/polito/mad/team19lab2/ui/OnSaleListFragment.kt
@@ -174,12 +174,12 @@ class OnSaleListFragment: Fragment(),SearchDialogFragment.NoticeDialogListener{
         var m1=minprice
         var m2=maxprice
         if(!minprice.isNullOrEmpty())
-            if(minprice.toInt()<=0)
+            if(minprice.toFloat()<=0)
                 m1=null
             else
                 search=true
         if(!maxprice.isNullOrEmpty())
-            if(maxprice.toInt()>=999999)
+            if(maxprice.toFloat()>=999999)
                 m2=null
             else
                 search=true
